+++ backend/tests/util.py
@@ -629,23 +629,24 @@ class LoRATestCase(_BaseTestCase):
 
 
 class ConfigTestCase(LoRATestCase):
-    """Testcase with configuration database support.
-
-    """
+    """Testcase with configuration database support."""
 
     def set_global_conf(self, conf):
-        conf_db.set_global_conf(conf)
+        conf_db.set_configuration({'org_units': dict(conf)})
 
     @classmethod
     def setUpClass(cls):
-        conf_db.testdb_setup()
+        conf_db.config["configuration"]["database"]["name"] = "test_confdb"
         super().setUpClass()
 
     @classmethod
     def tearDownClass(cls):
-        conf_db.testdb_teardown()
         super().tearDownClass()
 
     def setUp(self):
-        conf_db.testdb_reset()
+        conf_db._createdb(force=False)
         super().setUp()
+
+    def tearDown(self):
+        conf_db.drop_db()
+        super().tearDown()