+++ frontend/src/api/Employee.js
@@ -7,13 +7,6 @@ import store from '@/store'
 
 export default {
 
-  history (uuid) {
-    return Service.get(`/e/${uuid}/history/`)
-      .then(response => {
-        return response.data
-      })
-  },
-
   /**
    * Get engagement details for employee
    * @param {String} uuid - employee uuid