+++ frontend/src/views/employee/EmployeeDetail.vue
@@ -7,15 +7,11 @@ SPDX-License-Identifier: MPL-2.0
 
       <h4 class="card-title" v-show="!isLoading">
         <icon name="user-alt"/>
-        {{employee.name}} <span class="cpr">({{employee.cpr_no | CPRNumber}})</span>
+        {{employee.name}} <span class="cpr" v-if="showCPR">({{employee.cpr_no | CPRNumber}})</span>
       </h4>
 
       <div class="row">
         <div class="col"></div>
-
-        <div class="mr-3">
-          <mo-history :uuid="route.params.uuid" type="EMPLOYEE"/>
-        </div>
       </div>
 
       <employee-detail-tabs
@@ -35,7 +31,6 @@ SPDX-License-Identifier: MPL-2.0
 import '@/filters/CPRNumber'
 import { EventBus, Events } from '@/EventBus'
 import EmployeeDetailTabs from './EmployeeDetailTabs'
-import MoHistory from '@/components/MoHistory'
 import MoLoader from '@/components/atoms/MoLoader'
 import { mapState, mapGetters } from 'vuex'
 import { Employee } from '@/store/actions/employee'
@@ -43,7 +38,6 @@ import { Employee } from '@/store/actions/employee'
 export default {
   components: {
     EmployeeDetailTabs,
-    MoHistory,
     MoLoader
   },
 
@@ -66,7 +60,13 @@ export default {
     ...mapGetters({
       employee: Employee.getters.GET_EMPLOYEE,
       employeeDetails: Employee.getters.GET_DETAILS
-    })
+    }),
+
+    showCPR () {
+    let conf = this.$store.getters['conf/GET_CONF_DB']
+    return conf.show_cpr_no
+  },
+
   },
 
   created () {
@@ -79,6 +79,9 @@ export default {
   beforeDestroy () {
     EventBus.$off(Events.EMPLOYEE_CHANGED, this.listener)
   },
+
+
+
   methods: {
     loadContent (event) {
       this.latestEvent = event